% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{diagnosR-package}
\alias{diagnosR}
\alias{diagnosR-package}
\title{diagnosR: Quick Diagnostic Panel for LM}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

One-call diagnostic plots for linear models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ian-123/diagnosR}
  \item Report bugs at \url{https://github.com/Ian-123/diagnosR/issues}
}

}
\author{
\strong{Maintainer}: Ian Njuguna \email{njugunaian001@gmail.com}

}
\keyword{internal}
